tools/requirements.txt
python-dotenv
streamlit
openai

